+++ dpws/src/main/java/org/somda/sdc/dpws/soap/TransportInfo.java
@@ -24,7 +24,9 @@ public class TransportInfo {
     private final Integer localPort;
     private final String remoteAddress;
     private final Integer remotePort;
-    private List<X509Certificate> x509Certificates;
+
+    private final List<X509Certificate> x509Certificates;
+    private final String remoteNodeInfo;
 
     @Deprecated(since = "1.1.0", forRemoval = true)
     public TransportInfo(String scheme,
@@ -41,6 +43,7 @@ public class TransportInfo {
         // convert provided collection into a list
         this.x509Certificates = new ArrayList<>();
         this.x509Certificates.addAll(x509Certificates);
+        this.remoteNodeInfo = generateRemoteNodeInfo();
     }
 
     public TransportInfo(String scheme,
@@ -55,6 +58,14 @@ public class TransportInfo {
         this.remoteAddress = remoteAddress;
         this.remotePort = remotePort;
         this.x509Certificates = x509Certificates;
+        this.remoteNodeInfo = generateRemoteNodeInfo();
+    }
+
+    private String generateRemoteNodeInfo() {
+        return String.format("%s://%s:%s",
+                scheme,
+                remoteAddress != null ? remoteAddress : "<unknown-addr>",
+                remotePort != null ? remotePort : "<unknown-port>");
     }
 
     /**
@@ -109,5 +120,14 @@ public class TransportInfo {
      */
     public List<X509Certificate> getX509Certificates() {
         return x509Certificates;
+    }
+
+    /**
+     * Returns information of the remote node intended to be used for logging purposes.
+     *
+     * @return string representation comprising scheme, address and port.
+     */
+    public String getRemoteNodeInfo() {
+        return remoteNodeInfo;
     }
 }