+++ CHANGELOG.md
@@ -24,9 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `org.somda.sdc.biceps.common.storage.MdibStorage` method added to retrieve a list of states of a specific type. (#195)
 - `org.somda.sdc.biceps.consumer.preprocessing.DuplicateContextStateHandleHandler` to detect duplicate context state handles in MdibStateModifications. (#196)
 - `org.somda.sdc.biceps.consumer.access.RemoteMdibAccessImpl` and `org.somda.sdc.biceps.provider.access.LocalMdibAccessImpl` can be configured to specify which DescriptionPreprocessingSegments and StatePreprocessingSegments should be used for consumer or provider. (#196)
--  `org.somda.sdc.dpws.DpwsConfig#COMMUNICATION_LOG_WITH_HTTP_REQUEST_RESPONSE_ID`: configuration parameter to enable/disable generating HTTP request response transaction identifiers in communication logs. (#203)
--  `org.somda.sdc.dpws.DpwsConfig#COMMUNICATION_LOG_PRETTY_PRINT_XML`: configuration parameter to enable/disable pretty printing in communication logs. (#203)
-
+- `org.somda.sdc.dpws.DpwsConfig#COMMUNICATION_LOG_WITH_HTTP_REQUEST_RESPONSE_ID`: configuration parameter to enable/disable generating HTTP request response transaction identifiers in communication logs. (#203)
+- `org.somda.sdc.dpws.DpwsConfig#COMMUNICATION_LOG_PRETTY_PRINT_XML`: configuration parameter to enable/disable pretty printing in communication logs. (#203)
+- `org.somda.sdc.dpws.soap.getRemoteNodeInfo()`to retrieve a remote node's requested scheme, address and port. (#208)
+  
 ### Changed
 
 - Use of `io.github.threetenjaxb.core.LocalDateTimeXmlAdapter` to `org.somda.sdc.common.util.AnyDateTimeAdapter` for any XML Schema DateTime in module `biceps-model`. (#151)
@@ -45,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `org.somda.sdc.dpws.soap.HttpApplicationInfo` additional requestUri added, to determine the used POST address. (#190)
 - `org.somda.sdc.glue.provider.sco.ScoController` to process lists independent of a specific list type as otherwise 
   activate operations do not integrate well. (#207) 
+- `org.somda.sdc.dpws.soap.interception.RequestResponseObject` return non-optional `CommunicationContext` instances. (#208)
   
 ### Removed